refactor: moved the routes.ts file to constants
